test: consignments
