+++ tests/.gitignore
@@ -1,1 +1,3 @@
 /data/*.contract
+/data/*.rgb
+/data/storage*
